requirements.txt
apache-airflow
azure-cli
bs4
# pandas
# pyspark
# matplotlib
# seaborn
# nltk
# Flask